Add an ActivityUtils class and refactor some activities
