+++ app/src/main/java/ch/epfl/sweng/erpa/activities/LoginActivity.java
@@ -1,15 +1,10 @@
 package ch.epfl.sweng.erpa.activities;
 
-import android.app.Activity;
-import android.app.AlertDialog;
 import android.app.Application;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
-import android.widget.TextView;
 
 import com.annimon.stream.Optional;
 
@@ -20,6 +15,8 @@ import ch.epfl.sweng.erpa.model.UserAuth;
 import ch.epfl.sweng.erpa.services.UserAuthService;
 import toothpick.Scope;
 import toothpick.Toothpick;
+
+import static ch.epfl.sweng.erpa.util.ActivityUtils.createPopup;
 
 public class LoginActivity extends DependencyConfigurationAgnosticActivity {
     @Inject UserAuthService uas;
@@ -38,12 +35,12 @@ public class LoginActivity extends DependencyConfigurationAgnosticActivity {
         String passwordText = ((EditText) findViewById(R.id.password)).getText().toString();
 
         if (usernameText.isEmpty()) {
-            createPopup(getString(R.string.noNameMessage));
+            createPopup(getString(R.string.noNameMessage), this);
             return;
         }
 
         if (passwordText.isEmpty()) {
-            createPopup(getString(R.string.noPassMessage));
+            createPopup(getString(R.string.noPassMessage), this);
             return;
         }
 
@@ -51,34 +48,11 @@ public class LoginActivity extends DependencyConfigurationAgnosticActivity {
 
         ua.ifPresent(u -> finish());
 
-        createPopup(getString(R.string.incorrectLogin));
+        createPopup(getString(R.string.incorrectLogin), this);
     }
 
     public void continueWithoutLogin(View view) {
         Intent intent = new Intent(this, MainActivity.class);
         startActivity(intent);
-    }
-
-    private void createPopup(String text) {
-        AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
-
-        final TextView tv = new TextView(this);
-        tv.setText(text);
-        tv.setTextColor(Color.RED);
-        tv.setTextSize(16);
-
-        // set prompts.xml to alertdialog builder
-        alertDialogBuilder.setView(tv);
-
-        // set dialog message
-        alertDialogBuilder.setCancelable(false).setPositiveButton("OK", new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int id) {
-            }
-        });
-
-        // create alert dialog
-        AlertDialog alertDialog = alertDialogBuilder.create();
-        // show it
-        alertDialog.show();
     }
 }
