+++ app/src/main/java/ch/epfl/sweng/erpa/util/ActivityUtils.java
@@ -1,0 +1,33 @@
+package ch.epfl.sweng.erpa.util;
+
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.graphics.Color;
+import android.widget.TextView;
+
+public class ActivityUtils {
+
+    public static void createPopup(String text, Context context) {
+        AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(context);
+
+        final TextView tv = new TextView(context);
+        tv.setText(text);
+        tv.setTextColor(Color.RED);
+        tv.setTextSize(16);
+
+        // set prompts.xml to alertdialog builder
+        alertDialogBuilder.setView(tv);
+
+        // set dialog message
+        alertDialogBuilder.setCancelable(false).setPositiveButton("OK", new DialogInterface.OnClickListener() {
+            public void onClick(DialogInterface dialog, int id) {
+            }
+        });
+
+        // create alert dialog
+        AlertDialog alertDialog = alertDialogBuilder.create();
+        // show it
+        alertDialog.show();
+    }
+}
